backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
playwright
pdfplumber
transformers
torch
supabase
python-dotenv
earthengine-api
pandas
